Add comments, more cleanup
