+++ app/controllers/admin/settings_controller.rb
@@ -1,4 +1,7 @@
 module Admin
+  # This controller is solely responsible for rendering the settings page at
+  # /admin/customization/config. The actual updates get handled by the settings
+  # controllers in the Admin::Settings namespace.
   class SettingsController < Admin::ApplicationController
     layout "admin"
 
@@ -9,8 +12,8 @@ module Admin
 
     private
 
-    def authorization_resource
-      nil
-    end
+    # We need to override this method from Admin::ApplicationController since
+    # there is no resource to authorize.
+    def authorization_resource; end
   end
 end
